Whole bunches of changes for testing mostly.
Created code to use US census data to randomly generate test files with
names and numbers in it.  Updated the relevant tests once this was done as well.

Added some file and content checks to the iMISFile.read() method, it will throw an
error now if it finds no data in the file.